Crear un solo usuario mejorado

+++ scripts/crear-usuario.py
@@ -13,7 +13,6 @@ print()
 print("--- Datos del usuario --------------")
 usuario = click.prompt("Nombre de usuario", default="test")
 contrasenya = click.prompt("Contraseña", hide_input=True, confirmation_prompt="Confirmar contraseña")
-grupo = click.prompt("Grupo de usuarios", default=usuario)
 
 print("--- Cuota de recursos --------------")
 cpus = click.prompt("CPUs", type=float, default=4)
@@ -37,10 +36,6 @@ else:
     inicial_privadas = click.prompt("Número de host inicial para las direcciones privadas", default=10)
 
 print()
-
-print(f"Creando el grupo {grupo}...")
-os.system("onegroup create " + grupo)
-os.system("onevdc delgroup default " + grupo)
 
 print(f"\nCreando el usuario {usuario} y sus recursos asociados...\n")
 
@@ -69,7 +64,6 @@ os.system("onevdc adddatastore " + usuario + " 0 2")
 print(f"Creando el usuario {usuario}...")
 os.system("oneuser create " + usuario + " " + contrasenya)
 os.system("oneuser chgrp " + usuario + " " + usuario)
-os.system("oneuser addgroup " + usuario + " " + grupo)
 
 f = open("temp.txt", "w")
 f.write(f'''
@@ -90,37 +84,38 @@ os.system("oneuser quota " + usuario + " temp.txt")
 os.remove("temp.txt")
 
 try:
-    print(f"Creando la red {usuario}...")
-    f = open("temp.txt", "w")
-    f.write(f"NAME = \"{usuario}\"" + "\n")
-    f.write("BRIDGE = \"aulas\"" + "\n")
-    f.write("BRIDGE_TYPE = \"linux\"" + "\n")
-    f.write(f"DNS = \"{dns}\"" + "\n")
-    f.write(f"GATEWAY = \"{gateway}\"" + "\n")
-    f.write(f"NETWORK_MASK = \"{mascara}\"" + "\n")
-    f.write("OUTER_VLAN_ID = \"\"" + "\n")
-    f.write("PHYDEV = \"\"" + "\n")
-    f.write("SECURITY_GROUPS = \"0\"" + "\n")
-    f.write("VLAN_ID = \"\"" + "\n")
-    f.write("VN_MAD = \"fw\"" + "\n")
-    if publicas > 0:
-        f.write("AR=[TYPE = \"IP4\", IP = \""
-                + subred[:-1] + str(inicial_publicas) + "\", SIZE = \""
-                + str(publicas) + "\" ]" + "\n")
-        if privadas > 0:
+    if publicas > 0 or privadas > 0:
+        print(f"Creando la red {usuario}...")
+        f = open("temp.txt", "w")
+        f.write(f"NAME = \"{usuario}\"" + "\n")
+        f.write("BRIDGE = \"aulas\"" + "\n")
+        f.write("BRIDGE_TYPE = \"linux\"" + "\n")
+        f.write(f"DNS = \"{dns}\"" + "\n")
+        f.write(f"GATEWAY = \"{gateway}\"" + "\n")
+        f.write(f"NETWORK_MASK = \"{mascara}\"" + "\n")
+        f.write("OUTER_VLAN_ID = \"\"" + "\n")
+        f.write("PHYDEV = \"\"" + "\n")
+        f.write("SECURITY_GROUPS = \"0\"" + "\n")
+        f.write("VLAN_ID = \"\"" + "\n")
+        f.write("VN_MAD = \"fw\"" + "\n")
+        if publicas > 0:
+            f.write("AR=[TYPE = \"IP4\", IP = \""
+                    + subred[:-1] + str(inicial_publicas) + "\", SIZE = \""
+                    + str(publicas) + "\" ]" + "\n")
+            if privadas > 0:
+                f.write("AR=[TYPE = \"IP4\", IP = \""
+                        + subred[:-1] + str(inicial_privadas) + "\", SIZE = \""
+                        + str(privadas) + "\" ]" + "\n")
+        elif privadas > 0:
             f.write("AR=[TYPE = \"IP4\", IP = \""
                     + subred[:-1] + str(inicial_privadas) + "\", SIZE = \""
                     + str(privadas) + "\" ]" + "\n")
-    elif privadas > 0:
-        f.write("AR=[TYPE = \"IP4\", IP = \""
-                + subred[:-1] + str(inicial_privadas) + "\", SIZE = \""
-                + str(privadas) + "\" ]" + "\n")
-    f.close()
-    vnet_id = subprocess.run(['onevnet', 'create', 'temp.txt'],
-                             capture_output=True, text=True).stdout.split()[-1]
-    print(f"ID: {vnet_id}")
-    os.remove("temp.txt")
+        f.close()
+        vnet_id = subprocess.run(['onevnet', 'create', 'temp.txt'],
+                                 capture_output=True, text=True).stdout.split()[-1]
+        print(f"ID: {vnet_id}")
+        os.remove("temp.txt")
 
-    os.system("onevdc addvnet " + usuario + " 0 " + vnet_id)
+        os.system("onevdc addvnet " + usuario + " 0 " + vnet_id)
 except:
     print("Error al crear la red.")
